fix(react-ui-form): address PR review feedback
- Fix sort logic: indexOf returns -1 not undefined, use rank helper
- Break circular import: extract SettingsItem to own module
- Remove type re-exports from Settings.tsx (export from index.ts instead)
- Preserve numeric literal values in select onChange handler
- Fix number field clearing: empty string yields undefined not 0
- Fix literal filter/index mismatch: use flatMap instead of filter+map
